refactor: split local and AI managers
